ad smth

+++ shop/src/main/java/com/houseplant/shop/ground/repository/GroundRepository.java
@@ -26,7 +26,7 @@ public interface GroundRepository extends JpaRepository<Ground, Long> {
     @Transactional
     @Modifying
     @Query(value = "UPDATE ground SET name = :name, type = :type, price=:price, description = :description, stockQuantity = :stockQuantity, imageUrl = :imageUrl WHERE id = :id", nativeQuery = true)
-    void updateGround(final String name, final GroundType type,final Float price, final String description,  final int stockQuantity,  final byte[] imageUrl, final Long id);
+    void updateGround(final String name, final GroundType type,final double price, final String description,  final int stockQuantity,  final byte[] imageUrl, final Long id);
 
     @Transactional
     @Modifying
